+++ changelog.md
@@ -1,14 +1,19 @@
 Release Log:
 
+a1.2.3
+- Main menu options screen
+
+a1.2.2
+- Scroll to zoom
+- Gravity & physics adjustments
+
 a1.2.1
 - Physics adjustments & Bug fixes
-- Scroll to zoom
 
 a1.2_01 (labeled as 1.1.1_01)
 - Bug fixes
 
 a1.2 (labeled as a1.1.1)
-* Major additions&reworks
 - New pixel art set
 - Changed hex art
 - Added a spin attack